# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yunora AI Question Generation Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /boards:
    get:
      operationId: listBoards
      tags: [boards]
      summary: List all boards
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of boards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardList"
    post:
      operationId: createBoard
      tags: [boards]
      summary: Create a board
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardInput"
      responses:
        "201":
          description: Created board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
  /boards/{id}:
    get:
      operationId: getBoard
      tags: [boards]
      summary: Get board by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
    patch:
      operationId: updateBoard
      tags: [boards]
      summary: Update board
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardUpdate"
      responses:
        "200":
          description: Updated board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
    delete:
      operationId: deleteBoard
      tags: [boards]
      summary: Delete board
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── STANDARDS ────────────────────────────────────────────────────────────────
...
  /ai-providers:
    get:
      operationId: listAiProviders
      tags: [aiProviders]
      summary: List AI providers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiProvider"
    post:
      operationId: createAiProvider
      tags: [aiProviders]
      summary: Add a new AI provider
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiProviderInput"
      responses:
        "201":
          description: Created provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProvider"
  /ai-providers/{id}:
    patch:
      operationId: updateAiProvider
      tags: [aiProviders]
      summary: Update AI provider
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiProviderUpdate"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProvider"
    delete:
      operationId: deleteAiProvider
      tags: [aiProviders]
      summary: Delete AI provider
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /ai-providers/{id}/test:
    post:
      operationId: testAiProvider
      tags: [aiProviders]
      summary: Test an AI provider connection
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderTestResult"

  # ── PAPERS ────────────────────────────────────────────────────────────────────
...
  /analytics/dashboard:
    get:
      operationId: getDashboardStats
      tags: [analytics]
      summary: Get dashboard summary statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /analytics/questions-by-subject:
    get:
      operationId: getQuestionsBySubject
      tags: [analytics]
      summary: Questions grouped by subject
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Breakdown by subject
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubjectCount"
  /analytics/questions-by-difficulty:
    get:
      operationId: getQuestionsByDifficulty
      tags: [analytics]
      summary: Questions grouped by difficulty
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Breakdown by difficulty
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DifficultyCount"
  /analytics/model-usage:
    get:
      operationId: getModelUsage
      tags: [analytics]
      summary: AI model usage statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Model usage stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelUsage"
  /analytics/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [analytics]
      summary: Recent generation activity
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /analytics/monthly-report:
    get:
      operationId: getMonthlyReport
      tags: [analytics]
      summary: Monthly generation report (last 12 months)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyReport"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Board:
      type: object
      required: [id, name, code]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        description: { type: ["string", "null"] }
        isActive: { type: boolean }
        standardsCount: { type: integer }
        createdAt: { type: string, format: date-time }
    BoardInput:
      type: object
      required: [name, code]
      properties:
        name: { type: string }
        code: { type: string }
        description: { type: string }
        isActive: { type: boolean, default: true }
    BoardUpdate:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
        description: { type: string }
        isActive: { type: boolean }
    BoardList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Board"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── STANDARDS ─────────────────────────────────────────────────────────────
...
    AiProvider:
      type: object
      required: [id, name, providerType, defaultModel, isActive]
      properties:
        id: { type: integer }
        name: { type: string }
        providerType: { type: string }
        defaultModel: { type: string }
        availableModels:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    AiProviderInput:
      type: object
      required: [name, providerType, accessToken, defaultModel]
      properties:
        name: { type: string }
        providerType: { type: string, enum: [github_models, openai, azure_openai, anthropic, gemini, groq] }
        accessToken: { type: string }
        defaultModel: { type: string }
        availableModels:
          type: array
          items: { type: string }
        isActive: { type: boolean, default: true }
    AiProviderUpdate:
      type: object
      properties:
        name: { type: string }
        accessToken: { type: string }
        defaultModel: { type: string }
        availableModels:
          type: array
          items: { type: string }
        isActive: { type: boolean }
    ProviderTestResult:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string }
        latencyMs: { type: ["integer", "null"] }

    # ── PAPERS ────────────────────────────────────────────────────────────────
...
    DashboardStats:
      type: object
      required: [totalQuestions, totalSubjects, totalChapters, totalTopics, questionsToday, questionsThisMonth, totalBoards, totalStandards, activeProviders, totalPapers, avgQualityScore]
      properties:
        totalQuestions: { type: integer }
        totalSubjects: { type: integer }
        totalChapters: { type: integer }
        totalTopics: { type: integer }
        questionsToday: { type: integer }
        questionsThisMonth: { type: integer }
        totalBoards: { type: integer }
        totalStandards: { type: integer }
        activeProviders: { type: integer }
        totalPapers: { type: integer }
        avgQualityScore: { type: number }
        totalJobsRunning: { type: integer }
    SubjectCount:
      type: object
      required: [subjectName, count]
      properties:
        subjectName: { type: string }
        count: { type: integer }
    DifficultyCount:
      type: object
      required: [difficulty, count]
      properties:
        difficulty: { type: string }
        count: { type: integer }
    ModelUsage:
      type: object
      required: [model, count, avgQualityScore]
      properties:
        model: { type: string }
        provider: { type: string }
        count: { type: integer }
        avgQualityScore: { type: number }
    ActivityItem:
      type: object
      required: [id, jobId, action, description, createdAt]
      properties:
        id: { type: integer }
        jobId: { type: string }
        action: { type: string }
        description: { type: string }
        questionsGenerated: { type: ["integer", "null"] }
        model: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    MonthlyReport:
      type: object
      required: [month, year, count, avgQualityScore]
      properties:
        month: { type: integer }
        year: { type: integer }
        label: { type: string }
        count: { type: integer }
        avgQualityScore: { type: number }